v0.1: Minified + Fixed Webview + Format Selection
