+++ .vscodeignore
@@ -15,3 +15,4 @@ vsc-extension-quickstart.md
 .github/**
 assets/*
 !assets/icon.webp
+.prettierignore
